Profile View and Create Account Fixes
Profile view now accurately displays the last 5 stations visited by the user, and a navigation button was added to each entry of the profile. Additionally, the login screen and create account screen had their default values replaced with hint text. Also added some setters/getters for the username and userID to the MainActivity